examples/mcp/poc/requirements.txt
# External Tools Adapter PoC Requirements
eggai>=0.1.0
httpx>=0.25.0
mcp>=1.0.0
dspy-ai>=2.4.0
aioconsole>=0.6.0
python-dotenv>=1.0.0